requirements.txt
DateTime==4.3
GitPython==2.1.11
PyDriller==1.5.1
PyYAML==3.13
